+++ src/components/menu/examples/menu-grid.tsx
@@ -38,39 +38,53 @@ export class MenuGridExample {
     private items: Array<MenuItem | ListSeparator> = [
         {
             text: 'Companies',
-            icon: 'organization',
-            iconColor: 'rgb(var(--color-blue-default)',
+            icon: {
+                name: 'organization',
+                color: 'rgb(var(--color-blue-default)',
+            },
         },
         {
             text: 'Deals',
-            icon: 'money',
-            iconColor: 'rgb(var(--color-green-default))',
+            icon: {
+                name: 'money',
+                color: 'rgb(var(--color-green-default))',
+            },
         },
         {
             text: 'Agreements',
-            icon: 'handshake',
-            iconColor: 'rgb(var(--color-pink-default))',
+            icon: {
+                name: 'handshake',
+                color: 'rgb(var(--color-pink-default))',
+            },
         },
         {
             text: 'Todos',
-            icon: 'today',
-            iconColor: 'rgb(var(--color-teal-default))',
+            icon: {
+                name: 'today',
+                color: 'rgb(var(--color-teal-default))',
+            },
         },
         {
             text: 'History',
-            icon: 'comments',
-            iconColor: 'rgb(var(--color-grey-light))',
+            icon: {
+                name: 'comments',
+                color: 'rgb(var(--color-grey-light))',
+            },
         },
         { separator: true },
         {
             text: 'Coworkers',
-            icon: 'gender_neutral_user',
-            iconColor: 'rgb(var(--color-orange-light))',
+            icon: {
+                name: 'gender_neutral_user',
+                color: 'rgb(var(--color-orange-light))',
+            },
         },
         {
             text: 'Persons',
-            icon: 'user_group_man_man',
-            iconColor: 'rgb(var(--color-yellow-dark)',
+            icon: {
+                name: 'user_group_man_man',
+                color: 'rgb(var(--color-yellow-dark)',
+            },
         },
     ];