+++ src/components/picker/examples/picker-icons.tsx
@@ -18,96 +18,122 @@ export class PickerIconsExample {
             secondaryText:
                 'Anchor Hook, Anchor Drop, Ink Spray, Ink Propulsion',
             value: 1,
-            icon: 'octopus',
-            iconColor: 'var(--lime-red)',
+            icon: {
+                name: 'octopus',
+                color: 'var(--lime-red)',
+            },
         },
         {
             text: 'Ayla',
             secondaryText: 'Evil Eye, Rage, Chain Whack, Hop Skip',
             value: 2,
-            icon: 'visible',
-            iconColor: 'var(--lime-magenta)',
+            icon: {
+                name: 'visible',
+                color: 'var(--lime-magenta)',
+            },
         },
         {
             text: 'Clunk',
             secondaryText: 'Vacuum Bite, Explode, Missiles, Jet Boost',
             value: 3,
-            icon: 'robot_3',
-            iconColor: 'var(--lime-dark-blue)',
+            icon: {
+                name: 'robot_3',
+                color: 'var(--lime-dark-blue)',
+            },
         },
         {
             text: 'Coco',
             secondaryText: 'Ball Lightning, Blaze, Shock, Ollie',
             value: 4,
-            icon: 'surfing',
-            iconColor: 'var(--lime-blue)',
+            icon: {
+                name: 'surfing',
+                color: 'var(--lime-blue)',
+            },
         },
         {
             text: 'Derpl',
             secondaryText: 'Grid Trap, Siege Mode, Cat Shot, Booster Rocket',
             value: 5,
-            icon: 'cat',
-            iconColor: 'var(--lime-green)',
+            icon: {
+                name: 'cat',
+                color: 'var(--lime-green)',
+            },
         },
         {
             text: 'Froggy G',
             secondaryText:
                 'Splash Dash, Tornado Move, Bolt .45 Fish-gun, Frog Jump',
             value: 6,
-            icon: 'frog',
-            iconColor: 'var(--lime-turquoise)',
+            icon: {
+                name: 'frog',
+                color: 'var(--lime-turquoise)',
+            },
         },
         {
             text: 'Gnaw',
             secondaryText: 'Acid Spit, Grow Weedling, Bite, Skroggle Jump',
             value: 7,
-            icon: 'dog',
-            iconColor: 'var(--lime-orange)',
+            icon: {
+                name: 'dog',
+                color: 'var(--lime-orange)',
+            },
         },
         {
             text: 'Lonestar',
             secondaryText:
                 'Dynamite Throw, Summon Hyper Bull, Blaster, Double Jump',
             value: 8,
-            icon: 'sheriff',
-            iconColor: 'var(--lime-deep-red)',
+            icon: {
+                name: 'sheriff',
+                color: 'var(--lime-deep-red)',
+            },
         },
         {
             text: 'Leon',
             secondaryText: 'Tounge Snatch, Cloaking Skin, Slash, Reptile Jump',
             value: 9,
-            icon: 'croissant',
-            iconColor: 'var(--lime-yellow)',
+            icon: {
+                name: 'croissant',
+                color: 'var(--lime-yellow)',
+            },
         },
         {
             text: 'Raelynn',
             secondaryText:
                 'Timerift, Snipe, Protoblaster, Six Million Solar Human Jump',
             value: 10,
-            icon: 'sniper_rifle',
-            iconColor: 'var(--lime-dark-grey)',
+            icon: {
+                name: 'sniper_rifle',
+                color: 'var(--lime-dark-grey)',
+            },
         },
         {
             text: 'Skølldir',
             secondaryText: 'Mighty Throw, Earthquake, Bash, Explosive Fart',
             value: 11,
-            icon: 'beer',
-            iconColor: 'var(--lime-orange)',
+            icon: {
+                name: 'beer',
+                color: 'var(--lime-orange)',
+            },
         },
         {
             text: 'Voltar',
             secondaryText:
                 'Suicide Drones, Healbot, Techno Synaptic Wave, Hover',
             value: 12,
-            icon: 'brain',
-            iconColor: 'var(--lime-magenta)',
+            icon: {
+                name: 'brain',
+                color: 'var(--lime-magenta)',
+            },
         },
         {
             text: 'Yuri',
             secondaryText: 'Mine Deploying, Time Warp, Laser, Jet Pack',
             value: 13,
-            icon: 'year_of_monkey',
-            iconColor: 'var(--lime-light-grey)',
+            icon: {
+                name: 'year_of_monkey',
+                color: 'var(--lime-light-grey)',
+            },
         },
     ];