+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# [37.1.0-next.58](https://github.com/Lundalogik/lime-elements/compare/v37.1.0-next.57...v37.1.0-next.58) (2023-11-10)
+
+
+### Bug Fixes
+
+
+* **file-viewer:** use fileType to create a URL for the file ([26c8b5b](https://github.com/Lundalogik/lime-elements/commit/26c8b5b7f0e9f1ac052007655a49eddad2c25aaa))
+
+# [37.1.0-next.57](https://github.com/Lundalogik/lime-elements/compare/v37.1.0-next.56...v37.1.0-next.57) (2023-11-10)
+
+
+### Bug Fixes
+
+
+* **breadcrumbs:** support both `icon: string` & `icon: {name: string}` ([34753cf](https://github.com/Lundalogik/lime-elements/commit/34753cfeaf2ac789bc0782ba72c2076c7a70b98c))
+
+# [37.1.0-next.56](https://github.com/Lundalogik/lime-elements/compare/v37.1.0-next.55...v37.1.0-next.56) (2023-11-10)
+
+
+### Bug Fixes
+
+
+* **file-viewer:** ensure PDFs are loaded properly ([194738c](https://github.com/Lundalogik/lime-elements/commit/194738c23c601f3195a4c25da10d70840158e57d))
+
 # [37.1.0-next.55](https://github.com/Lundalogik/lime-elements/compare/v37.1.0-next.54...v37.1.0-next.55) (2023-11-08)